Check if upload directory exists before trying to remove it when deleting profile

+++ app/Http/Controllers/ProfileController.php
@@ -61,7 +61,9 @@ class ProfileController extends Controller
             $file->delete();
         }
 
-        rmdir($path);
+        if (is_dir($path)) {
+            rmdir($path);
+        }
 
         Auth::logout();
